requirements.txt
fastapi
uvicorn[standard]

python-dotenv
motor
email-validator
PyJWT
bcrypt
python-multipart
python-jose

httpx
pydantic

langchain
langchain-community
langgraph
langsmith

langchain-google-genai
langchain-gemini

PyPDF2
pymupdf
matplotlib
gradio
